+++ datatable/java/datatable/src/main/java/io/cucumber/datatable/DataTable.java
@@ -88,7 +88,7 @@ public final class DataTable {
      * @throws IllegalArgumentException when the table is not rectangular or contains null values
      */
     public static DataTable create(List<List<String>> raw, TableConverter tableConverter) {
-        return new DataTable(copy(requireNonNullEntries(requireRectangularTable(raw))), tableConverter);
+        return new DataTable(copy(requireRectangularTable(raw)), tableConverter);
     }
 
     private static List<List<String>> copy(List<List<String>> balanced) {
@@ -104,21 +104,6 @@ public final class DataTable {
             rawCopy.add(unmodifiableList(rowCopy));
         }
         return unmodifiableList(rawCopy);
-    }
-
-    private static List<List<String>> requireNonNullEntries(List<List<String>> raw) {
-        // Iterate in case of linked list.
-        int rowIndex = 0;
-        for (List<String> row : raw) {
-            int columnIndex = row.indexOf(null);
-            if (columnIndex >= 0) {
-                throw new IllegalArgumentException(
-                        "raw contained null at row: " + rowIndex + " column: " + columnIndex);
-            }
-            rowIndex++;
-        }
-
-        return raw;
     }
 
     private static List<List<String>> requireRectangularTable(List<List<String>> table) {